moved new function to different folder in lib directory
